added new pics
